chore: export report from ui
