+++ packages/ui/package.json
@@ -15,6 +15,10 @@
       "types": "./dist/index.d.ts",
       "import": "./dist/index.js"
     },
+    "./reporter": {
+      "types": "./dist/reporter.d.ts",
+      "import": "./dist/reporter.js"
+    },
     "./*": "./*"
   },
   "main": "./dist/index.js",
@@ -38,7 +42,9 @@
     "prepublishOnly": "pnpm build"
   },
   "dependencies": {
-    "sirv": "^2.0.2"
+    "sirv": "^2.0.2",
+    "fast-glob": "^3.2.12",
+    "flatted": "^3.2.7"
   },
   "devDependencies": {
     "@faker-js/faker": "^7.6.0",
@@ -59,7 +65,6 @@
     "cypress": "^11.2.0",
     "d3-graph-controller": "^2.4.0",
     "diff": "^5.1.0",
-    "flatted": "^3.2.7",
     "floating-vue": "^2.0.0-y.0",
     "picocolors": "^1.0.0",
     "rollup": "^2.79.1",
