+++ packages/vitest/src/node/reporters/index.ts
@@ -5,7 +5,6 @@ import { VerboseReporter } from './verbose'
 import { TapReporter } from './tap'
 import { JUnitReporter } from './junit'
 import { TapFlatReporter } from './tap-flat'
-import { HTMLReporter } from './html'
 
 export { DefaultReporter }
 
@@ -17,7 +16,6 @@ export const ReportersMap = {
   'tap': TapReporter,
   'tap-flat': TapFlatReporter,
   'junit': JUnitReporter,
-  'html': HTMLReporter,
 }
 
 export type BuiltinReporters = keyof typeof ReportersMap
